% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transX.R
\name{ID2Anno}
\alias{ID2Anno}
\title{link the targeted ID to annotation}
\usage{
ID2Anno(abbrevID, annotation, outName = "matchIDAnno")
}
\arguments{
\item{abbrevID}{A data.frame file with '.csv' format including the compounds ID.}

\item{annotation}{A data.frame file with '.csv' format annotated information for ID. 
The 1st column is the abbrevID.}

\item{outName}{An output table with ".xls".}
}
\description{
ID2Anno is used to link the compounds ID to annotation.
}
\author{
Hemi Luan, hemi.luan@gmail.com
}
