% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTargetGUI_v2.R
\name{statTargetGUI}
\alias{statTargetGUI}
\title{statTargetGUI for statTarget software}
\usage{
statTargetGUI()
}
\value{
The output of GUI
}
\description{
The statTarget GUI session. The Signal Correction and 
Statistical Analysis session are included in statTarget 2.0 software. 
See the details at https://stattarget.github.io
}
\examples{
if (interactive()) {statTargetGUI()}
}
\references{
Dunn WB., et al. Nat Protoc. 2011, 6, pp1060.
Luan H., et al. GigaScience 2015, 4, pp16.
Luan H., et al. J. Proteome Res., 2015, 14, pp467.
}
\author{
Hemi Luan hemi.luan@gmail.com
}
\keyword{A}
\keyword{Ensemble}
\keyword{GUI}
\keyword{Metabolomics}
\keyword{Statistical}
\keyword{analysis}
\keyword{correction}
\keyword{for}
\keyword{learning}
\keyword{of}
\keyword{signal}
\keyword{statTarget}
