% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transX.R
\name{transX}
\alias{transX}
\title{transX for statTarget inputs}
\usage{
transX(data, type)
}
\arguments{
\item{data}{A transX objects. The output file from Mass Spectrometry Data 
softwares.}

\item{type}{The output file formats from Mass Spectrometry Data software, 
including 'XCMS' or 'xcms','SIEVE' or 'sieve' and 'skyline' or 'SKYLINE';}
}
\value{
An output directory named 'statTargetDirectory'
}
\description{
transX is to generate statTarget input file formats from 
Mass Spectrometry Data softwares, such as XCMS, SIEVE and SKYLINE
}
\examples{
datpath <- system.file('extdata',package = 'statTarget')
dataXcms <- paste(datpath,'xcmsOutput.tsv', sep='/')
dataSkyline <- paste(datpath,'skylineDemo.csv', sep='/')
transX(dataXcms,'xcms')
}
\author{
Hemi Luan, hemi.luan@gmail.com
}
\keyword{SIEVE}
\keyword{SKYLINE}
\keyword{XCMS}
\keyword{inputs}
